Refactored Contact into __base to fix circular import on tests.
