+++ src/pysdmx/model/__base.py
@@ -3,8 +3,6 @@ from typing import Iterator, Optional, Sequence
 
 from msgspec import Struct
 
-from pysdmx.model import Contact
-
 
 class Annotation(Struct, frozen=True, omit_defaults=True):
     """Annotation class.
@@ -164,7 +162,7 @@ class Agency(MaintainableArtefact, frozen=True, omit_defaults=True):
         contacts: The contact of the agency.
     """
 
-    contacts = Sequence[Contact]
+    contacts = Sequence["Contact"]
 
 
 class Item(NameableArtefact, frozen=True, omit_defaults=True):
@@ -213,3 +211,33 @@ class ItemScheme(MaintainableArtefact, frozen=True, omit_defaults=True):
             return None
         else:
             return out[0]
+
+
+class Contact(Struct, frozen=True, omit_defaults=True):
+    """Contact details such as the name of a contact and his email address.
+
+    Attributes:
+        id: An identifier for a contact. If the contact represents a person,
+            this could be the person's username in the organisation.
+        name: The contact name, which could be the name of a person, the name
+            of a service ("e.g. Support"), etc.
+        department: The department in which the contact is located (e.g.
+            "Statistics").
+        role: The contact's role, which could be his job title, or a role such
+            as data owner, data steward, subject matter expert, etc.
+        telephones: A list of telephone numbers.
+        faxes: A list of fax numbers.
+        uris: A list of URLs relevant for the contact (e.g. a link to an online
+            form that can be used to send questions, a link to a support forum,
+            etc.).
+        emails: a list of email addresses.
+    """
+
+    id: Optional[str] = None
+    name: Optional[str] = None
+    department: Optional[str] = None
+    role: Optional[str] = None
+    telephones: Optional[Sequence[str]] = None
+    faxes: Optional[Sequence[str]] = None
+    uris: Optional[Sequence[str]] = None
+    emails: Optional[Sequence[str]] = None
