+++ src/pysdmx/model/__init__.py
@@ -7,6 +7,7 @@ subset** of the SDMX information model.
 from re import Pattern
 from typing import Any
 
+from pysdmx.model.__base import Contact, Item, ItemScheme
 from pysdmx.model.category import Category, CategoryScheme
 from pysdmx.model.code import (
     Code,
@@ -37,7 +38,7 @@ from pysdmx.model.map import (
     ValueMap,
 )
 from pysdmx.model.metadata import MetadataAttribute, MetadataReport
-from pysdmx.model.organisation import Contact, DataflowRef, Organisation
+from pysdmx.model.organisation import DataflowRef, Organisation
 
 
 def encoders(obj: Any) -> Any:
@@ -97,6 +98,8 @@ __all__ = [
     "Hierarchy",
     "HierarchyAssociation",
     "ImplicitComponentMap",
+    "Item",
+    "ItemScheme",
     "StructureMap",
     "MetadataAttribute",
     "MetadataReport",
